[monitored-universe][4] Universe implementation in AID

+++ aim/agent/aid/universes/aim_universe.py
@@ -20,6 +20,7 @@ from oslo_log import log as logging
 
 from aim.agent.aid.universes.aci import converter
 from aim.agent.aid.universes import base_universe as base
+from aim.api import resource as aim_resource
 from aim.api import status as aim_status
 from aim import context
 from aim.db import tree_model
@@ -42,8 +43,13 @@ class AimDbUniverse(base.HashTreeStoredUniverse):
         self.tree_manager = tree_model.TenantHashTreeManager()
         self.context = context.AimContext(db_session)
         self._converter = converter.AciToAimModelConverter()
+        self._converter_aim_to_aci = converter.AimToAciModelConverter()
         self._served_tenants = set()
         return self
+
+    @property
+    def name(self):
+        return "AIM_Config_Universe"
 
     def serve(self, tenants):
         LOG.debug('Serving tenants: %s' % tenants)
@@ -93,7 +99,12 @@ class AimDbUniverse(base.HashTreeStoredUniverse):
             if dissected[0] == ACI_FAULT:
                 fault_code = dissected[1][-1]
                 dissected = self._dissect_key(key[:-1])
-            klass = converter.resource_map[dissected[0]][0]['resource']
+                key = key[:-1]
+            aci_dn = apic_client.DNManager().build(
+                [tuple(x.split('|')) for x in key])
+            aci_klass = apic_client.DNManager().aci_decompose_dn_guess(
+                aci_dn, dissected[0])[0]
+            klass = converter.resource_map[aci_klass][0]['resource']
             res = klass(
                 **dict([(y, dissected[1][x])
                         for x, y in enumerate(klass.identity_attributes)]))
@@ -129,40 +140,57 @@ class AimDbUniverse(base.HashTreeStoredUniverse):
         return self.get_resources(resource_keys)
 
     def push_resources(self, resources):
+        return self._push_resources(resources, monitored=False)
+
+    def _push_resources(self, resources, monitored=False):
         fault_method = {'create': self.manager.set_fault,
                         'delete': self.manager.clear_fault}
         for method in resources:
-            for item in resources[method]:
+            if method == 'delete':
+                # Use ACI items directly
+                items = resources[method]
+            else:
+                # Convert everything before creating
+                items = self._converter.convert(resources[method])
+            for resource in items:
                 # Items are in the other universe's format unless deletion
                 try:
-                    if method == 'delete':
-                        converted = [item]
+                    if isinstance(resource, aim_status.AciFault):
+                        # Retrieve fault's parent and set/unset the fault
+                        parent = self._retrieve_fault_parent(resource)
+                        LOG.debug(
+                            "%s for object %s: %s" %
+                            (fault_method[method].__name__,
+                             parent.__dict__, resource.__dict__))
+                        fault_method[method](self.context,
+                                             resource=parent,
+                                             fault=resource)
                     else:
-                        converted = self._converter.convert([item])
-                    for resource in converted:
-                        if isinstance(resource, aim_status.AciFault):
-                            # Retrieve fault's parent and set/unset the fault
-                            parent = self._retrieve_fault_parent(resource)
-                            LOG.debug(
-                                "%s for object %s: %s" %
-                                (fault_method[method].__name__,
-                                 parent.__dict__, resource.__dict__))
-                            fault_method[method](self.context,
-                                                 resource=parent,
-                                                 fault=resource)
+                        LOG.debug(
+                            "%s object %s" %
+                            (fault_method[method].__name__,
+                             resource.__dict__))
+                        if isinstance(resource,
+                                      aim_resource.Tenant) and monitored:
+                            # Monitored Universe doesn't interact with Tenant
+                            # Resources
+                            continue
+                        if method == 'create':
+                            if monitored:
+                                # We need two more conversions to screen out
+                                # unmanaged items
+                                resource.monitored = monitored
+                                resource = self._converter_aim_to_aci.convert(
+                                    [resource])
+                                resource = self._converter.convert(resource)[0]
+                                resource.monitored = monitored
+                            self.manager.create(self.context, resource,
+                                                overwrite=True)
                         else:
-                            LOG.debug(
-                                "%s object %s" %
-                                (fault_method[method].__name__,
-                                 resource.__dict__))
-                            if method == 'create':
-                                self.manager.create(self.context, resource,
-                                                    overwrite=True)
-                            else:
-                                self.manager.delete(self.context, resource)
+                            self.manager.delete(self.context, resource)
                 except Exception as e:
                     LOG.error("Failed to %s object %s in AIM: %s." %
-                              (method, item, e.message))
+                              (method, resource, e.message))
                     LOG.debug(traceback.format_exc())
 
     def _retrieve_fault_parent(self, fault):
@@ -183,6 +211,39 @@ class AimDbOperationalUniverse(AimDbUniverse):
     def state(self):
         return self._get_state(tree=tree_model.OPERATIONAL_TREE)
 
+    @property
+    def name(self):
+        return "AIM_Operational_Universe"
+
     def get_optimized_state(self, other_state):
         return super(AimDbOperationalUniverse, self).get_optimized_state(
             other_state, tree=tree_model.OPERATIONAL_TREE)
+
+    def reconcile(self, other_universe, delete_candidates):
+        # When the other universes are ok with deleting a Tenant, there's no
+        # reason for the Operational Universe to oppose that decision
+        return self._reconcile(other_universe, delete_candidates,
+                               always_vote_deletion=True)
+
+
+class AimDbMonitoredUniverse(AimDbUniverse):
+    @property
+    def state(self):
+        return self._get_state(tree=tree_model.MONITORED_TREE)
+
+    @property
+    def name(self):
+        return "AIM_Monitored_Universe"
+
+    def get_optimized_state(self, other_state):
+        return super(AimDbMonitoredUniverse, self).get_optimized_state(
+            other_state, tree=tree_model.MONITORED_TREE)
+
+    def push_resources(self, resources):
+        self._push_resources(resources, monitored=True)
+
+    def reconcile(self, other_universe, delete_candidates):
+        # We want monitored universe to stop reconciling when the corresponding
+        # AIM tenant doesn't exist.
+        return self._reconcile(other_universe, delete_candidates,
+                               skip_dummy=True)
